+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Program "export-netspoc" exports visible services again.
+  This was still needed.
+
 ## [2026-03-02-1256]
 
 ### Fixed